requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
pypdf
langchain-huggingface
faiss-cpu
langchain_chroma
chromadb
arxiv
wikipedia
<<<<<<< HEAD
=======
validators==0.28.1
youtube_transcript_api
>>>>>>> 709625f (HealthAI)
